stashing
Changes so far:
- removed endpoint FirstAssemblyview
- moved the functionality to template_protein field in
TargetSerializer
- removed TargetMoleculesserializer
- removed sequences field from TargetSerializer

This is a result of Boris' comment in
github (https://github.com/m2ms/fragalysis-frontend/issues/1373#issuecomment-2047417958)
where he said the field isn't used and template_protein field is not
used. Looking at the code where this may be used, revealed that
Targetmoleculesserializer can be removed as well

NB! they're not removed-removed right now, only commented in. This
commit can be used to restore the code.