fix: resolve migration failure when merging several diverging issues
After migrations were on hold for a while, merging back several issues with migrations caused a migration error. This was caused by sligtly wrong order of procedures (possibly as a result of an automatic merge) but also an oversight on my part - as the compound identifiers are loaded from compounds_manual.csv during LHS data loading I didn't think to test with RHS data, but evidently this partly touches the same models and filled the tables with incompatible data types. This patch should resolve these issues without a data loss.