feat: enable loading assay data
Data loaded from .xlsx or .csv file. User needs to supply column name to use as an id. This is used to find objects from database to link the uploaded data to.
Accepts files with single-row header and double row header. In the latter, the second row denotes the data type. Currently, 'float' and 'text' are supported.