feat: major dependency upgrade
This is as far as I can take it now, need some input.

The following packages were not upgraded because they're pinned to specific or really old versions:
- im-squonk2-client = "^1.22.2"  (latest 6.1.0)
- bcrypt = "4.0.1"               (latest 5.0.0)
- django-prometheus = "2.3.1"    (latest 2.4.1)

Need to consult exactly why.

The following packages were not upgraded because they seem to be unused:
- deepdiff = "^6.2.0"
- graphene-django = "^3.1.3"
- mysql-connector-python = "^8.1.0"
- psutil = "^5.9.5"
- pydiscourse = "^1.3.0"

pydiscourse is currently used, but AFAIU, going to be dropped.

A curious case is the package django-webpack-loader which is held at version 0.7 while 3.2 is out. Noticed it's throwing errors on startup (although the stack seems to work). Upgrading this failed, stack won't deploy.